progress

+++ src/main/java/net/frozenblock/lib/gravity/mixin/PrimedTntMixin.java
@@ -20,6 +20,7 @@ package net.frozenblock.lib.gravity.mixin;
 
 import net.frozenblock.lib.gravity.impl.EntityGravityInterface;
 import net.minecraft.world.entity.item.PrimedTnt;
+import net.minecraft.world.phys.Vec3;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArgs;
@@ -30,9 +31,9 @@ public abstract class PrimedTntMixin implements EntityGravityInterface {
 
 	@ModifyArgs(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/phys/Vec3;add(DDD)Lnet/minecraft/world/phys/Vec3;", ordinal = 0))
 	private void useGravity(Args args) {
-		double x = (double) args.get(0);
+		double x = args.get(0);
 		double y = (double) args.get(1) + this.frozenLib$getGravity();
-		double z = (double) args.get(2);
+		double z = args.get(2);
 		Vec3 gravity = this.frozenLib$getEffectiveGravity();
 
 		args.set(0, x - gravity.x);
