+++ src/main/java/net/frozenblock/lib/gravity/mixin/ThrowableProjectileMixin.java
@@ -20,6 +20,7 @@ package net.frozenblock.lib.gravity.mixin;
 
 import net.frozenblock.lib.gravity.impl.EntityGravityInterface;
 import net.minecraft.world.entity.projectile.ThrowableProjectile;
+import net.minecraft.world.phys.Vec3;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
@@ -35,9 +36,9 @@ public abstract class ThrowableProjectileMixin implements EntityGravityInterface
 
 	@ModifyArgs(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/projectile/ThrowableProjectile;setDeltaMovement(DDD)V", ordinal = 0))
 	private void useGravity(Args args) {
-		double x = (double) args.get(0);
+		double x = args.get(0);
 		double y = (double) args.get(1) + this.frozenLib$getGravity();
-		double z = (double) args.get(2);
+		double z = args.get(2);
 		Vec3 gravity = this.frozenLib$getEffectiveGravity();
 
 		args.set(0, x - gravity.x);
